% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTripleMI.R
\name{checkTripleMI}
\alias{checkTripleMI}
\title{Check Consistency of Conditional Independence for a Triple of Nodes
used in FCI-MI for multiple imputed data sets of continuous data}
\usage{
checkTripleMI(
  a,
  b,
  c,
  nbrsA,
  nbrsC,
  data,
  sepsetA,
  sepsetC,
  alpha,
  version.unf = c(NA, NA),
  maj.rule = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{a, b, c}{(Integer) positions in adjacency matrix for nodes a, b, and c,
respectively.}

\item{nbrsA, nbrsC}{(Integer) position in adjacency matrix for neighbors of
a and c, respectively.}

\item{data}{An object of type mids, which stands for 'multiply imputed
data set', typically created by a call to function mice().}

\item{sepsetA}{Vector containing Sepset(a,c).}

\item{sepsetC}{Vector containing Sepset(c,a).}

\item{alpha}{significance level of test.}

\item{version.unf}{(Integer) vector of length two. See
[functioname(pcalg::chekTriple)] for more details.}

\item{maj.rule}{Logical indicating how the majority rule is applied. See
[functioname(pcalg::chekTriple)] for more details.}

\item{verbose}{Logical asking for detailed output of intermediate steps.}
}
\value{
See [functioname(pcalg::chekTriple)] for details.
}
\description{
This function is a modification of [functioname(pcalg::checkTriple)]
to be used for multiple imputation.
}
\note{
This is a modified function of [functioname(pcalg::pc)]
      from the package 'pcalg' (Kalisch et al., 2012;
      http://www.jstatsoft.org/v47/i11/).
}
