% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussCItestMI.R
\name{gaussCItestMI}
\alias{gaussCItestMI}
\title{Test Conditional Independence of Gaussians via Fisher's Z using
multiple imputations}
\usage{
gaussCItestMI(x, y, S, data)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S,
respectively, in the adjacency matrix.
It is tested, whether X and Y are conditionally independent
given the subset S of the remaining nodes.}

\item{data}{An object of type mids, which stands for 'multiply imputed
data set', typically created by a call to function mice()}
}
\value{
Returns the p-value of the test
}
\description{
This function is a modification of \code{pcalg::\link[pcalg:condIndFisherZ]{gaussCItest}}
to be used for multiple imputation.
}
\examples{
## load data (numeric variables)
dat <- as.matrix(windspeed)

## delete some observations
set.seed(123)
dat[sample(1:length(dat), 260)] <- NA

## Impute missing values under normal model
imp <- mice(dat, method = "norm")

## analyse data
# complete data:
suffStat <- list( C = cor(windspeed), n = nrow(windspeed) )
gaussCItest(1, 2, c(4,5), suffStat = suffStat)
# multiple imputation:
gaussCItestMI(1, 2, c(4,5), data = imp)
# test-wise deletion:
gaussCItwd(1, 2, c(4,5), suffStat = dat)
# list-wise deletion:
suffStat <- list( C = cor(dat[complete.cases(dat), ]), n = sum(complete.cases(dat)) )
gaussCItest(1, 2, c(4,5), suffStat = suffStat)

}
