% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussCItwd.R
\name{gaussCItwd}
\alias{gaussCItwd}
\title{Fisher's z-Test for (Conditional) Independence between Gaussian Variables with Missings}
\usage{
gaussCItwd(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in each correlation matrix in \code{suffStat}. It is tested 
whether X and Y are conditionally independent given the subset S of the 
remaining variables.}

\item{suffStat}{\code{data.frame} containing the raw data.}
}
\value{
See \code{\link[pcalg:condIndFisherZ]{gaussCItest}} for details on 
Fisher's z-test. Test-wise deletion is valid if missingness does not jointly 
depend on X and Y.

A p-value.
}
\description{
A wrapper for \code{\link[pcalg:condIndFisherZ]{gaussCItest}}, 
to be used within \code{\link[pcalg]{skeleton}}, \code{\link[pcalg]{pc}} or 
\code{\link[pcalg]{fci}} when the data contain missing values. Observations 
where at least one of the variables involved in the test is missing are 
deleted prior to performing the test (test-wise deletion).
}
\seealso{
\code{\link[pcalg]{gaussCItest}} for complete data, 
\code{\link[pcalg]{gaussMItest}} for multiply imputed data
}
