% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formulas_saturated.R
\name{make.formulas.saturated}
\alias{make.formulas.saturated}
\title{Creates a \code{formulas} Argument}
\usage{
make.formulas.saturated(
  data,
  blocks = mice::make.blocks(data),
  predictorMatrix = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} with the source data.}

\item{blocks}{An optional specification for blocks of variables in the rows. 
The default assigns each variable in its own block.}

\item{predictorMatrix}{A \code{predictorMatrix} specified by the user.}
}
\value{
A list of formulas.
}
\description{
This helper function creates a valid \code{formulas} object. 
The \code{formulas} object is an argument to the \code{\link[mice]{mice}} function. 
It is a list of formulas that specifies the target variables and the predictors 
by means of the standard ~ operator. In contrast to \code{\link[mice]{make.formulas}}, 
which creates main effects formulas, \code{make.formulas.saturated} 
creates formulas including all possible interaction effects.
}
\note{
A modification of \code{\link[mice]{make.formulas}} by Stef van Buuren et al.
}
\seealso{
\code{\link[mice]{make.formulas}}
}
