% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixCItest.R
\name{mixCItest}
\alias{mixCItest}
\title{Likelihood Ratio Test for (Conditional) Independence between Mixed Variables}
\usage{
mixCItest(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in \code{suffStat}. It is tested whether X and Y are conditionally 
independent given the subset S of the remaining variables.}

\item{suffStat}{\code{data.frame}. Discrete variables must be coded as factors.}
}
\value{
A p-value.
}
\description{
A likelihood ratio test for (conditional) independence between mixed 
(continuous and unordered categorical) variables, to be used within 
\link[pcalg]{skeleton}, \link[pcalg]{pc} or \link[pcalg]{fci}. 
It assumes that the variables in the test follow a Conditional Gaussian 
distribution, i.e. conditional on each combination of values of the discrete
 variables, the continuous variables are multivariate Gaussian. 
 Each multivariate Gaussian distribution is allowed to have its own mean 
 vector and covariance matrix.
}
\details{
The implementation follows Andrews et al. (2018). The same test is also 
implemented in TETRAD and in the R-package rcausal, a wrapper for the 
TETRAD Java library. Small differences in the p-values returned by 
CGtest and the TETRAD/rcausal equivalent are due to differences in 
handling sparse or empty cells, and differences in covariance estimation.

Note that the test assumes a Conditional Gaussian distribution 
(see Lauritzen & Wermuth, 1989) for the variables \emph{in that specific test}, 
rather than for all variables in the graph / data set. Due to the 
non-collapsibility of the Conditional Gaussian distribution, this implies 
that the assumptions of two tests with overlapping sets of variables may 
contradict each other. Little is known about how this affects the quality of 
the estimated graph in practice.
}
\references{
Andrews B., Ramsey J., Cooper G.F. (2018): Scoring Bayesian networks 
of mixed variables. \emph{International Journal of Data Science and Analytics} 6:3-18.

Lauritzen S.L., Wermuth N. (1989): Graphical models for associations between 
variables, some of which are qualitative and some quantitative. 
\emph{The Annals of Statistics} 17(1):31-57. 

Scheines R., Spirtes P., Glymour C., Meek C., Richardson T. (1998): 
The TETRAD project: Constraint based aids to causal model specification. 
\emph{Multivariate Behavioral Research} 33(1):65-117. 
http://www.phil.cmu.edu/tetrad/index.html
}
\author{
Janine Witte
}
