% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsepMI.R
\name{pdsepMI}
\alias{pdsepMI}
\title{Estimate Final Skeleton in the FCI-MI algorithm for multiple imputed
data sets of continuous data}
\usage{
pdsepMI(
  skel,
  data,
  p,
  sepset,
  alpha,
  pMax,
  m.max = Inf,
  pdsep.max = Inf,
  NAdelete = TRUE,
  unfVect = NULL,
  biCC = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{skel}{Graph object returned by [skeletonMI())]

[skeletonMI())]: R:skeletonMI())}

\item{data}{An object of type mids, which stands for 'multiply imputed
data set', typically created by a call to function mice()}

\item{p}{Number of variables.}

\item{sepset}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{alpha}{Significance level for the individual conditional independence tests.}

\item{pMax}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{m.max}{Maximum size of the conditioning sets that are considered in the conditional independence tests.}

\item{pdsep.max}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{NAdelete}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{unfVect}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{biCC}{See \code{pcalg::\link[pcalg]{pdsep}} for more details.}

\item{verbose}{Logical indicating that detailed output is to be provided.}
}
\value{
See \code{pcalg::\link[pcalg]{pdsep}} for more details.
}
\description{
This function is a modification of \code{\link[pcalg:pdsep]{pcalg::pdsep()}}
to be used for multiple imputation.
}
