% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.graph.R
\name{boot.graph}
\alias{boot.graph}
\title{Bootstrap Resampling for the PC-MI- and the FCI-MI-algorithm}
\usage{
boot.graph(data, method = c("pcMI", "fciMI"), args, R, m = 10,
  seed = NA, quickpred = FALSE, ...)
}
\arguments{
\item{data}{Data frame with missing values}

\item{method}{Character string specifying the algorithm for causal discovery
from the package 'pcalg'.}

\item{args}{Arguments passed to `method`. NOTE: argument `labels` is set
internally and should not be used!}

\item{R}{A positive integer number of bootstrap replications.}

\item{seed}{A positive integer that is used as argument for set.seed().}

\item{quickpred}{If true, mice uses quickpred to select predictors.}

\item{...}{Further arguments passed to the imputation function `mice()`.}
}
\value{
List of objects of class `pcalgo` (see [functionname(pcalg::`pcAlgo-class`)])
        or of `fcmialgo` (see [functionname(pcalg::`fciAlgo-class`)]).
}
\description{
Generate R bootstrap replicates for the PC or FCI algorithm for data with
missing values.
}
\examples{
library(mice)
daten <- windspeed[,1]
for(i in 2:ncol(windspeed)) daten <- c(daten, windspeed[,i])
daten[sample(1:length(daten), 260)] <- NA
daten <- matrix(daten, ncol = 6)

boot.graph <- boot.graph(
            data = daten,
            method = "pcMI",
            args = "solve.confl = TRUE, alpha = 0.05",
            R = 10)
}
