% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disCItwd.R
\name{disCItwd}
\alias{disCItwd}
\title{G square Test for (Conditional) Independence between Discrete Variables with Missings}
\usage{
disCItwd(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in \code{suffStat}. It is tested whether X and Y 
are conditionally independent given the subset S of the remaining variables.}

\item{suffStat}{a list with three elements, \code{"dm"}, \code{"nlev"}, 
\code{"adaptDF"}; each corresponding to the above arguments.}
}
\value{
A p-value.
}
\description{
A wrapper for \code{\link[pcalg]{disCItest}}, to be used within 
\code{\link[pcalg]{skeleton}}, \code{\link[pcalg]{pc}} or 
\code{\link[pcalg]{fci}} when the data contain missing values. 
Observations where at least one of the variables involved in the test is 
missing are deleted prior to performing the test (test-wise deletion).
}
\details{
See \code{\link{disCItest}} for details the G square test. Test-wise deletion 
is valid if missingness does not jointly depend on X and Y.
}
\seealso{
\code{\link[pcalg]{disCItest}} for complete data, \code{\link[pcalg]{disMItest}}
for multiply imputed data
}
