% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disCItwd.R
\name{disCItwd}
\alias{disCItwd}
\title{G square Test for (Conditional) Independence between Discrete Variables with Missings}
\usage{
disCItwd(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S,
respectively, in \code{suffStat}. It is tested whether X and Y
are conditionally independent given the subset S of the remaining variables.}

\item{suffStat}{a list with three elements, \code{"dm"}, \code{"nlev"},
\code{"adaptDF"}; each corresponding to the above arguments.}

\item{dm}{data matrix (rows: samples, columns: variables) with integer entries; the k levels for a given column must be coded by the integers 0,1,...,k-1. (see example)}

\item{nlev}{optional vector with numbers of levels for each variable in \code{dm}.}

\item{adaptDF}{logical specifying if the degrees of freedom should be lowered by one for each zero count. The value for the degrees of freedom cannot go below 1.}
}
\value{
A p-value.
}
\description{
A wrapper for \code{pcalg::\link[pcalg]{disCItest}}, to be used within
\code{pcalg::\link[pcalg]{skeleton}}, \code{pcalg::\link[pcalg]{pc}} or
\code{pcalg::\link[pcalg]{fci}} when the data contain missing values.
Observations where at least one of the variables involved in the test is
missing are deleted prior to performing the test (test-wise deletion).
}
\details{
See \code{\link[pcalg]{disCItest}} for details on the G square test. Test-wise deletion
is valid if missingness does not jointly depend on X and Y.
}
\examples{

## simulate variables (integers 0,...,k)
n <- 200
set.seed(123)
x <- sample(0:2, n, TRUE) # 3 levels
y <- sample(0:3, n, TRUE) # 4 levels
z <- sample(0:1, n, TRUE) # 2 levels
dat <- cbind(x,y,z)

## delete some observations of y and z
dat[sample(1:n, 40), 2] <- NA
dat[sample(1:n, 40), 3] <- NA

## analyse incomplete data
# test-wise deletion:
suffStat <- list(dm = dat, nlev = c(3,4,2), adaptDF = FALSE)
disCItwd(1, 3, NULL, suffStat = suffStat)
# list-wise deletion:
suffStat2 <- list(dm = dat[complete.cases(dat), ], nlev = c(3,4,2), adaptDF = FALSE)
disCItest(1, 3, NULL, suffStat = suffStat2)

}
\seealso{
\code{pcalg::\link[pcalg]{disCItest}} for complete data, \code{\link{disMItest}}
for multiply imputed data
}
