% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disMItest.R
\name{disMItest}
\alias{disMItest}
\title{G square Test for (Conditional) Independence between Discrete Variables after 
Multiple Imputation}
\usage{
disMItest(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in \code{suffStat}. It is tested whether X and Y are conditionally 
independent given the subset S of the remaining variables.}

\item{suffStat}{A list of \code{data.frame}s containing the multiply imputed 
data sets. Usually obtained from a \code{\link[mice:mids-class]{mice:mids}} 
object using \code{\link[mice:complete.mids]{complete}} with argument 
\code{action="all"}. All variables must be coded as factors.}
}
\value{
A p-value.
}
\description{
A version of \code{\link[pcalg]{disCItest}}, to be used within 
\code{\link[pcalg]{skeleton}}, \code{\link[pcalg]{pc}} or 
\code{\link[pcalg]{fci}} when multiply imputed data sets are available. 
Note that in contrast to \code{\link[pcalg]{disCItest}}, the variables must 
here be coded as factors.
}
\details{
See \code{\link{mixCItest}} for details on the assumptions of the 
Conditional Gaussian likelihood ratio test. CGtestMI applies this test to each 
\code{data.frame} in \code{suffStat}, then combines the results using the rules 
in Meng & Rubin (1992). Degrees of freedom are never adapted, and there is no 
minimum required sample size, while \code{\link[pcalg]{disCItest}} requires 
\code{10*df} observations and otherwise returns a p-value of 1.
}
\references{
Meng X.-L., Rubin D.B. (1992): Performing likelihood ratio tests with multiply 
imputed data sets. \emph{Biometrika} 79(1):103-111.
}
\seealso{
\code{\link[pcalg]{disCItest}} for complete data
}
\author{
Janine Witte
}
