% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexCItest.R
\name{flexCItest}
\alias{flexCItest}
\title{Wrapper for gaussMItest, disMItest and mixMItest}
\usage{
flexCItest(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S,
respectively, in the dataset. It is tested whether X and Y are conditionally
independent given the subset S of the remaining variables.}

\item{suffStat}{a list generated using \code{\link{getSuff}} with \code{test="flexMItest"}. See below for details.}
}
\value{
A p-value.
}
\description{
A plug-in conditional independence test for \code{pcalg::\link[pcalg]{skeleton}}, \code{pcalg::\link[pcalg]{pc}} or
\code{pcalg::\link[pcalg]{fci}} when multiply imputed data sets are available. \code{flexMItest} detects whether
variables are continuous, discrete or mixed, and automatically switches between \code{\link{gaussMItest}} (continuous only),
\code{link{disMItest}} (discrete only) and \code{\link{mixMItest}} (mixed).
}
\details{
\code{suffStat} needs to be a list with four elements named \code{datlist}, \code{corlist},
\code{conpos} and \code{dispos}. \code{datlist} is the list of imputed datasets. \code{corlist}
is a list with M+1 elements, where M is the number of imputed datasets. For i=1,...,M, the
the i-th element of \code{corlist} is the correlation matrix of the continuous variables in the i-th imputed dataset;
the (M+1)-the element is the number of rows in each imputed dataset.
\code{conpos} is a vector containing the integer positions of the continuous variables in the original dataset.
\code{dispos} is a vector containing the integer positions of the discrete variables in the original dataset.
}
\seealso{
\code{\link{gaussMItest}}, \code{\link{disMItest}} and \code{\link{mixMItest}}
}
