% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussCItestMI.R
\name{gaussCItestMI}
\alias{gaussCItestMI}
\title{Test Conditional Independence of Gaussians via Fisher's Z using
multiple imputations}
\usage{
gaussCItestMI(x, y, S, data)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S,
respectively, in the adjacency matrix.
It is tested, whether X and Y are conditionally independent
given the subset S of the remaining nodes.}

\item{data}{An object of type mids, which stands for 'multiply imputed
data set', typically created by a call to function mice()}
}
\value{
Returns the p-value of the test
}
\description{
This function is a modification of [functioname(pcalg::gaussCItest)]
to be used for multiple imputation.
}
\examples{
library(mice)
daten <- windspeed[,1]
for(i in 2:ncol(windspeed)) daten <- c(daten, windspeed[,i])
daten[sample(1:length(daten), 260)] <- NA
daten <- matrix(daten, ncol = 6)

## Impute missing values
imp <- mice(daten)

## analyse data
gaussCItestMI(1,2,c(4,5), imp)
gaussCItest(1,2,c(4,5), suffStat = list(C = cor(windspeed), n = nrow(windspeed)))
gaussCItest(1,2,c(4,5), suffStat = list(C = cor(daten[complete.cases(daten),]),
            n = nrow(daten[complete.cases(daten),])))

}
