% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixCItwd.R
\name{mixCItwd}
\alias{mixCItwd}
\title{Likelihood Ratio Test for (Conditional) Independence between Mixed Variables with Missings}
\usage{
mixCItwd(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in \code{suffStat}. It is tested whether X and Y are conditionally 
independent given the subset S of the remaining variables.}

\item{suffStat}{\code{data.frame}. Discrete variables must be coded as factors.}
}
\value{
A p-value.
}
\description{
A version of \code{\link{mixCItest}}, to be used within \code{\link[pcalg]{skeleton}}, 
\code{\link[pcalg]{pc}} or \code{\link[pcalg]{fci}} when the data contain missing values. 
Observations where at least one of the variables involved in the test is missing 
are deleted prior to performing the test (test-wise deletion).
}
\details{
See \code{\link{mixCItest}} for details on the assumptions of the 
Conditional Gaussian likelihood ratio test. Test-wise deletion is valid if 
missingness does not jointly depend on X and Y.
}
\seealso{
\code{\link{mixCItest}} for complete data, 
\code{\link{mixMItest}} for multiply imputed data
}
