% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixMItest.R
\name{mixMItest}
\alias{mixMItest}
\title{Likelihood Ratio Test for (Conditional) Independence between Mixed Variables 
after Multiple Imputation}
\usage{
mixMItest(x, y, S = NULL, suffStat)
}
\arguments{
\item{x, y, S}{(integer) position of variable X, Y and set of variables S, 
respectively, in \code{suffStat}. It is tested whether X and Y are 
conditionally independent given the subset S of the remaining variables.}

\item{suffStat}{A list of \code{data.frame}s containing the multiply 
imputed data sets. Usually obtained from a \code{\link[mice:mids-class]{mice:mids}} 
object using \code{\link[mice:complete.mids]{complete}} with argument \code{action="all"}. 
Discrete variables must be coded as factors.}
}
\value{
A p-value.
}
\description{
A version of \link{mixCItest}, to be used within \link[pcalg]{skeleton}, 
\link[pcalg]{pc} or \link[pcalg]{fci} when multiply imputed data sets are available.
}
\details{
See \code{\link{mixCItest}} for details on the assumptions of the 
Conditional Gaussian likelihood ratio test. \code{CGtestMI} applies this test 
to each \code{data.frame} in \code{suffStat}, then combines the results using 
the rules in Meng & Rubin (1992).
}
\references{
Meng X.-L., Rubin D.B. (1992): Performing likelihood ratio tests 
with multiply imputed data sets. \emph{Biometrika} 79(1):103-111.
}
\author{
Janine Witte
}
