% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.cons.internMI.R
\name{pc.cons.internMI}
\alias{pc.cons.internMI}
\title{Utility for conservative and majority rule in PC-MI and FCI-MI
applied for multiple imputed data sets of continuous data}
\usage{
pc.cons.internMI(sk, data, alpha, version.unf = c(NA, NA),
  maj.rule = FALSE, verbose = FALSE)
}
\arguments{
\item{sk}{A skeleton object as returned from [functioname(micd::skeletonMI)]}

\item{data}{An object of type mids, which stands for 'multiply imputed
data set', typically created by a call to function mice().}

\item{alpha}{Significance level for the conditional independence test.}

\item{version.unf}{(Integer) vector of length two. See
[functioname(pcalg::chekTriple)] for more details.}

\item{maj.rule}{Logical indicating how the majority rule is applied. See
[functioname(pcalg::chekTriple)] for more details.}

\item{verbose}{Logical asking for detailed output of intermediate steps.}
}
\value{
See [functioname(pcalg::pc.cons.intern)] for more details.
}
\description{
This function is a modification of [functioname(pcalg::pc.cons.intern)]
to be used for multiple imputation.
}
\note{
This is a modified function of [functioname(pcalg::pc)]
      from the package 'pcalg' (Kalisch et al., 2012;
      http://www.jstatsoft.org/v47/i11/).
}
